96. Handling API error response part1
Exception should be exceptional, we should not exception to control the flow of exection in handler.

Its better to send object to controller, as success or failure. From controller, based on success and failure we can send http response from controller.

Changes:
Result.cs: create new file, to share genaric result as success or failure with error